PYIC-1718: Rename the user-identity lambda
