+++ lambdas/build-user-identity/src/main/resources/log4j2.yaml
