+++ settings.gradle
@@ -8,7 +8,7 @@ include "lib",
 		"lambdas:evaluate-gpg45-scores",
 		"lambdas:build-debug-credential-data",
 		"lambdas:initialise-ipv-session",
-		"lambdas:useridentity",
+		"lambdas:build-user-identity",
 		"lambdas:process-journey-step",
 		"lambdas:build-cri-oauth-request",
 		"lambdas:credentialissuererror",
